feat: Implement quota management features in user detail and dashboard
- Updated user-detail.html to display cost usage instead of token usage.
- Added quota management section with configuration and bonus granting modals.
- Enhanced user listing in users.html to reflect new quota logic.
- Introduced management navigation link for quota management in main layout.

+++ src/main/java/io/github/samzhu/ledger/dto/api/SystemUsageResponse.java
@@ -33,5 +33,5 @@ public record SystemUsageResponse(
      * @param requestCount 請求次數
      * @param totalTokens 總 token 數
      */
-    public record TopItem(String id, String name, int requestCount, long totalTokens) {}
+    public record TopItem(String id, String name, long requestCount, long totalTokens) {}
 }
